% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTweets_Roxygen.R
\docType{package}
\name{CleanTweets}
\alias{CleanTweets}
\title{CleanTweets}
\description{
This package offers cleaning functions of tweets extract with rtweets
}
\details{
For more information use ? in front of function names
}
\author{
Jean-François Savard \email{Jean-Francois.Savard@enap.ca}
}
